Better error handling and README updates
The UI state should be set on `on*` messages on the client socket. This
was set directly upon taking UI action, but this does not guarantee that
the action will succeed or not.

Also, added a toast when server abnormally dissconnects. This was not
handled before.

Updated README.rst to correctly include the image.